src/requirements.txt
aiohttp                   3.12.15
aiosignal                 1.4.0
anyio                     4.8.0
appdirs                   1.4.4
appnope                   0.1.4
argon2-cffi               23.1.0
argon2-cffi-bindings      21.2.0
arrow                     1.3.0
asttokens                 3.0.0
async-lru                 2.0.4
attrs                     25.3.0
audiobook-dl              0.7.3
babel                     2.17.0
beautifulsoup4            4.13.3
bleach                    6.2.0
certifi                   2025.1.31
cffi                      1.17.1
charset-normalizer        3.4.1
comm                      0.2.2
contourpy                 1.3.1
cssselect                 1.3.0
cv                        1.0.0
cycler                    0.12.1
debugpy                   1.8.13
decorator                 5.2.1
defusedxml                0.7.1
executing                 2.2.0
fastjsonschema            2.21.1
filelock                  3.19.1
fonttools                 4.56.0
fqdn                      1.5.1
frozenlist                1.7.0
fsspec                    2025.7.0
h11                       0.14.0
hf-xet                    1.1.8
httpcore                  1.0.7
httpx                     0.28.1
huggingface-hub           0.34.4
idna                      3.10
imageio                   2.37.0
importlib_resources       6.5.2
iniconfig                 2.1.0
ipykernel                 6.29.5
ipython                   9.0.2
ipython_pygments_lexers   1.1.1
isoduration               20.11.0
jedi                      0.19.2
Jinja2                    3.1.6
joblib                    1.4.2
json5                     0.10.0
jsonpointer               3.0.0
jsonschema                4.23.0
jsonschema-specifications 2024.10.1
jupyter_client            8.6.3
jupyter_core              5.7.2
jupyter-events            0.12.0
jupyter-lsp               2.2.5
jupyter_server            2.15.0
jupyter_server_terminals  0.5.3
jupyterlab                4.3.5
jupyterlab_pygments       0.3.0
jupyterlab_server         2.27.3
kiwisolver                1.4.8
kmodes                    0.12.2
lazy_loader               0.4
lightning-utilities       0.15.2
lpips                     0.1.4
lxml                      5.4.0
m3u8                      6.0.0
markdown-it-py            3.0.0
MarkupSafe                3.0.2
matplotlib                3.10.0
matplotlib-inline         0.1.7
mdurl                     0.1.2
mistune                   3.1.2
mpmath                    1.3.0
multidict                 6.6.4
mutagen                   1.47.0
natsort                   8.4.0
nbclient                  0.10.2
nbconvert                 7.16.6
nbformat                  5.10.4
nest-asyncio              1.6.0
networkx                  3.5
notebook                  7.3.2
notebook_shim             0.2.4
numpy                     2.2.3
opencv-python             4.12.0.88
overrides                 7.7.0
packaging                 24.2
pandas                    2.2.3
pandocfilters             1.5.1
parso                     0.8.4
pexpect                   4.9.0
pillow                    11.1.0
pip                       25.2
platformdirs              4.3.6
plotly                    5.24.1
pluggy                    1.6.0
prometheus_client         0.21.1
prompt_toolkit            3.0.50
propcache                 0.3.2
psutil                    7.0.0
ptyprocess                0.7.0
pure_eval                 0.2.3
pycountry                 24.6.1
pycparser                 2.22
pycryptodome              3.23.0
Pygments                  2.19.1
pyparsing                 3.2.1
pytest                    8.4.1
python-dateutil           2.9.0.post0
python-json-logger        3.3.0
pytorch-lightning         2.5.3
pytz                      2025.1
PyYAML                    6.0.2
pyzmq                     26.3.0
referencing               0.36.2
requests                  2.32.3
rfc3339-validator         0.1.4
rfc3986-validator         0.1.1
rich                      14.0.0
rpds-py                   0.23.1
safetensors               0.6.2
scikit-image              0.25.2
scikit-learn              1.6.1
scipy                     1.15.2
seaborn                   0.13.2
Send2Trash                1.8.3
setuptools                76.0.0
six                       1.17.0
sniffio                   1.3.1
soupsieve                 2.6
stack-data                0.6.3
sympy                     1.13.1
tenacity                  9.0.0
terminado                 0.18.1
threadpoolctl             3.6.0
tifffile                  2025.6.11
timm                      0.9.10
tinycss2                  1.4.0
tomli                     2.2.1
torch                     2.6.0
torchaudio                2.6.0
torchmetrics              1.8.1
torchvision               0.21.0
tornado                   6.4.2
tqdm                      4.67.1
traitlets                 5.14.3
types-python-dateutil     2.9.0.20241206
typing_extensions         4.12.2
tzdata                    2025.1
uri-template              1.3.0
urllib3                   2.3.0
wcwidth                   0.2.13
webcolors                 24.11.1
webencodings              0.5.1
websocket-client          1.8.0
yarl                      1.20.1